Synchronize with backend service updates
- data: network: sync API changes
- domain: sync db schema changes
- ui: sync API changes for UX

+++ app/src/main/java/com/mnnit/moticlubs/data/data_source/ClubDao.kt
@@ -17,7 +17,7 @@ interface ClubDao {
     suspend fun getClubs(): List<Club>
 
     @Query("SELECT * FROM club WHERE club.cid = :clubID")
-    suspend fun getClub(clubID: Int): Club
+    suspend fun getClub(clubID: Long): Club
 
     @Delete
     suspend fun deleteClub(club: Club)
